重构基本完成，项目能启动，但存在小bug

+++ src/models/order/order_log.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer, String
+from sqlalchemy import BigInteger, Column, ForeignKey, String
 from sqlalchemy.orm import relationship
 
 from src.models.base import BaseModel, RemarkMixin, SoftDeleteMixin, TimestampMixin
@@ -8,10 +8,10 @@ class OrderLog(BaseModel, TimestampMixin, SoftDeleteMixin, RemarkMixin):
     """订单操作日志模型"""
     __tablename__ = "order_log"
 
-    order_id = Column(Integer, ForeignKey("order.id"), nullable=False, index=True, comment="订单ID")
+    order_id = Column(BigInteger, ForeignKey("order.id"), nullable=False, index=True, comment="订单ID")
 
     operator_type = Column(String(20), default="user", nullable=False, index=True, comment="操作人类型：system/user/admin")
-    operator_id = Column(Integer, nullable=True, index=True, comment="操作人ID")
+    operator_id = Column(BigInteger, nullable=True, index=True, comment="操作人ID")
     operator_name = Column(String(50), nullable=True, comment="操作人姓名（冗余）")
 
     action = Column(String(50), nullable=False, index=True, comment="操作动作：create/pay/cancel/refund/expire/extend/upgrade")
