+++ src/models/order/order_refund.py
@@ -1,4 +1,4 @@
-from sqlalchemy import JSON, Column, DateTime, ForeignKey, Integer, String
+from sqlalchemy import JSON, BigInteger, Column, DateTime, ForeignKey, Integer, String
 from sqlalchemy.orm import relationship
 
 from src.models.base import BaseModel, RemarkMixin, SoftDeleteMixin, TimestampMixin
@@ -8,8 +8,8 @@ class OrderRefund(BaseModel, TimestampMixin, SoftDeleteMixin, RemarkMixin):
     """订单退款记录模型"""
     __tablename__ = "order_refund"
 
-    order_id = Column(Integer, ForeignKey("order.id"), nullable=False, index=True, comment="订单ID")
-    order_payment_id = Column(Integer, ForeignKey("order_payment.id"), nullable=True, index=True, comment="支付记录ID")
+    order_id = Column(BigInteger, ForeignKey("order.id"), nullable=False, index=True, comment="订单ID")
+    order_payment_id = Column(BigInteger, ForeignKey("order_payment.id"), nullable=True, index=True, comment="支付记录ID")
 
     refund_no = Column(String(50), nullable=True, unique=True, index=True, comment="退款单号")
     refund_method = Column(String(20), nullable=False, comment="退款方式：original/alipay/wechat/other")
@@ -19,7 +19,7 @@ class OrderRefund(BaseModel, TimestampMixin, SoftDeleteMixin, RemarkMixin):
     reason = Column(String(200), nullable=True, comment="退款原因")
 
     operator_type = Column(String(20), default="user", nullable=False, index=True, comment="操作人类型：system/user/admin")
-    operator_id = Column(Integer, nullable=True, index=True, comment="操作人ID")
+    operator_id = Column(BigInteger, nullable=True, index=True, comment="操作人ID")
     operator_name = Column(String(50), nullable=True, comment="操作人姓名（冗余）")
 
     refunded_at = Column(DateTime(timezone=True), nullable=True, comment="退款成功时间")
